backend/requirements.txt
fastapi
uvicorn
pymupdf
sentence-transformers
chromadb
python-dotenv
requests
edge-tts
nest_asyncio
python-multipart
motor
pymongo>=4.0
cloudinary
configobj
configparser
gTTS
httplib2
nipype
pandas
pyxnat
